Fetch random photos from Unsplash API working and displaying with pin button
