+++ src/App.js
@@ -7,36 +7,34 @@ import { BrowserRouter } from "react-router-dom";
 import RouteList from "./RouteList";
 
 function App() {
-  const [images, setImages] = useState([]);
-  const [userImages, setUserImages] = useState(
-    localStorage.getItem("userImages") || []
-  );
-
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(function fetchImagesOnLoad() {
-    async function fetchImages() {
-      axios.get("https://api.unsplash.com/photos/random?count=30&client_id=h-JqYpjiPbIbpYbTW9E3yluBUrgU1tCjb2ysGQqM-O0")
-            .then(response=>{
-              setImages(response.data)
-            })
-      setIsLoading(false);
-    }
-    fetchImages();
-  }, []);
-
-  function handlePinClick(src) {
-    //setstate....
-  }
+  //const [images, setImages] = useState([]);
+  //const [userImages, setUserImages] = useState(
+  //  localStorage.getItem("userImages") || []
+  //);
+//
+  //const [isLoading, setIsLoading] = useState(true);
+//
+  //useEffect(function fetchImagesOnLoad() {
+  //  async function fetchImages() {
+  //    axios.get("https://api.unsplash.com/photos/random?count=30&client_id=h-JqYpjiPbIbpYbTW9E3yluBUrgU1tCjb2ysGQqM-O0")
+  //          .then(response=>{
+  //            setImages(response.data)
+  //          })
+  //    setIsLoading(false);
+  //  }
+  //  fetchImages();
+  //}, []);
+//
+  //function handlePinClick(src) {
+  //  localStorage.setItem("userImages", [src, ...userImages]);
+  //}
 
   return (
     <div className="App">
       <BrowserRouter>
-        <NavBar>
           <div>
             <RouteList></RouteList>
           </div>
-        </NavBar>
       </BrowserRouter>
     </div>
   );
